switch to only storing 32 bits of the 64-bit hashes in the ldict

+++ src/binding.hpp
@@ -40,7 +40,7 @@ public:
     typedef uint64_t key_type;
     typedef uint64_t value_type;
     // pbf message cache
-    typedef google::sparse_hash_set<uint64_t> ldictcache;
+    typedef google::sparse_hash_set<uint32_t> ldictcache;
     // list + map as simple LRU cache
     typedef std::pair<std::string,std::string> message_pair;
     typedef std::list<message_pair> message_list;
